requirements.txt
keras==2.2.4 
Tensorflow-gpu==1.15.4
cv2
numpy
requests
tqdm
uuid
paho-mqtt
Pillow
ImageHash
